feat: detect iframe in react
