+++ docusaurus.config.js
@@ -39,10 +39,6 @@ const config = {
       src: "/js/redirects.js",
       async: false,
     },
-    {
-      src: "/js/iframe-detector.js",
-      async: false,
-    }
   ],
   presets: [
     [
